refactor: move DCAF to model file
