+++ models/deepctr/pnn.py
@@ -434,7 +434,7 @@ def fit_and_predict(
 
         submission["row_id"] = submission["row_id"] - 4000000
         # evaluate
-        filepath = f"./notebooks/jhkim/submission/pnn-{i+45}.csv"
+        filepath = f"./pnn-{i+45}.csv"
         filepaths.append(filepath)
         submission[["row_id", "is_clicked", "is_installed"]].to_csv(
             "./pnn.csv", index=False, sep="\t"
@@ -465,9 +465,6 @@ if __name__ == "__main__":
         import os
 
         os.makedirs(f"submissions/{run_id}", exist_ok=True)
-
-        # train = pd.read_parquet("featured_train.parquet")
-        # test = pd.read_parquet("featured_test.parquet")
 
         fit_and_predict(
             train, test, linear_feature_columns, dnn_feature_columns, mode=mode
